typescript `src/popover` (#488)
